Enable the CSS image-orientation always
The spec has been updated to something we can commit to.
Intent to Ship has been approved.

So implement the CSS image-orientation property for all
images (content and style) and change the valid values
to "from-image" or "none". Update tests accordingly.

This does not remove the flag in case we need to turn it
off.

Intent to Ship:
https://groups.google.com/a/chromium.org/forum/#!searchin/blink-dev/image-orientation%7Csort:date/blink-dev/jPgnRoivOis/X1E6SuOoAwAJ

Bug: 158753
Change-Id: Ib50d263b01f7cf6b2d17174138fd223a8105da8e