[OT-PW] Make it work with OOP-R
This CL makes off-thread paint worklet work with OOP-R. OOP-R serialize
an PaintOp and stores the serialized buffer in a shared memory. GPU
then reads the memory and de-serialize it.

For off-thread paint worklet, we are producing a DrawImageRectOp. Note
that this DrawImageRectOp is different than a normal one. It is a
DrawImageRectOp that contains a DrawRecord that contains multiple
PaintOps in it. So this CL adds a special code path to serialize it,
for paint worklet DrawImageRectOp only. The way to serialize it is to
iterate over all the PaintOps inside the DrawRecord. In this way, we
do not need to change anything on the GPU de-serialize code.

We also changed the layout tests.
1. We changed some tests, now using fillRect instead of strokeRect.
The problem with strokeRect is that using GPU rasterization could
easily result in subpixel difference on different platforms. Using
fillRect will avoid that.
2. There are some other tests (8 of them) that produces different
results compared software vs gpu rasterization. That's why we have
a oop-r and a non-oop-r under css-paint-api/ and csspaint/. The
oop-r contains tests that produces exactly the same result with and
without OOP-R. The non-oop-r tests will run without OOP-R, and these
tests are moved to virtual/ test dir and added expected.png when
running with OOP-R.

Bug: 1020238
Change-Id: Ie0aadfe217e462547cc16a273d09ec8a64f055a0